Refactor UI components and update todo list
A significant refactoring of the UI components in the application has been made to improve the modularity and hence the maintainability of the code. Specifically, three UI components were created: ManualBarcode, ManualPrice, and PaymentCash, which eliminates repeating code. Additionally, a number of tasks went updated on the todo list, helping keep the project organized and on track.