Fix payment and UI components, update to-do list
This commit refines the PaymentCash component to enhance the display of balance. It also amends layout elements within the _index. Two items were removed from the to-do list. The Admin component now triggers a customer card clearance upon emptying the cart, and a check for a non-null manual price has been added to the ManualPrice component.