Add product removal feature to cart
Product removal feature is implemented to the cart allowing users to decrease or completely remove the quantity of a specific product. This feature is accessible through a new button in the routes file connected to a 'removeProduct' function in the 'useCart' hook. The icon for the new feature was also added to Button component.