Modify button component and admin catalog page
Button component has been updated to adjust the width and height of the button in different screen sizes. Additional icons were added as well. On the admin catalog page, a functionality was implemented to let users hide or show the keyboard. The appearance of the page was also adjusted to accommodate these changes.